+++ Runtime/Core/CyclopsNext.cs
@@ -25,7 +25,7 @@ namespace Smonch.CyclopsFramework
 
         internal static CyclopsNext Rent(ICyclopsRoutineScheduler scheduler)
         {
-            var next = GenericPool<CyclopsNext>.Get();
+            CyclopsNext next = GenericPool<CyclopsNext>.Get();
 
             next.Scheduler = scheduler;
 
@@ -61,9 +61,8 @@ namespace Smonch.CyclopsFramework
         }
 
         public CyclopsLambda Loop(Action f)
-            => (CyclopsLambda)Add(CyclopsLambda.Instantiate(period: 0f, maxCycles: float.MaxValue, f));
+            => Add(CyclopsLambda.Instantiate(period: 0f, maxCycles: float.MaxValue, f));
         
-
         public CyclopsLambda Loop(float period, float maxCycles, Action f)
             => (CyclopsLambda)Add(CyclopsLambda.Instantiate(period, maxCycles, f));
 
@@ -77,14 +76,14 @@ namespace Smonch.CyclopsFramework
                     context!.Stop();
             }));
 
-            return (CyclopsLambda)routine;
+            return routine;
         }
 
         public CyclopsLambda LoopWhile(Func<bool> whilePredicate, Action whileBody, float period = 0f)
         {
             CyclopsLambda context = null;
             
-            var routine = context = Add(CyclopsLambda.Instantiate(period, float.MaxValue, () =>
+            CyclopsLambda routine = context = Add(CyclopsLambda.Instantiate(period, float.MaxValue, () =>
             {
                 if (whilePredicate())
                 {
@@ -99,12 +98,12 @@ namespace Smonch.CyclopsFramework
                 }
             }));
 
-            return (CyclopsLambda)routine;
+            return routine;
         }
 
         public CyclopsRoutine Nop(string tag = null, int maxCycles = 1)
         {
-            var nop = CyclopsNop.Instantiate(maxCycles);
+            CyclopsNop nop = CyclopsNop.Instantiate(maxCycles);
 
             if (tag != null)
                 nop.AddTag(tag);
@@ -137,34 +136,10 @@ namespace Smonch.CyclopsFramework
 
         public CyclopsUpdate Lerp(double period, double maxCycles, Action<float> f, Func<float, float> bias = null)
             => Add(CyclopsUpdate.Instantiate(period, maxCycles, bias, f));
-
-        //public CyclopsWaitForMessage ProcessAnalytics(string tag, Action<CyclopsMessage> f, int maxCycles = 1, float timeout = float.MaxValue)
-        //{
-        //    var seq = Listen(tag, Message_Analytics, timeout, maxCycles);
-        //    seq.OnSuccess(f);
-
-        //    return seq;
-        //}
-
-        //public CyclopsWaitForMessage WaitForAnalytics(string tag, int maxCycles = 1, float timeout = float.MaxValue)
-        //{
-        //    return Listen(tag, Message_Analytics, timeout, maxCycles);
-        //}
-
-        //public CyclopsWaitForMessage WaitForAnalytics(CyclopsRequirement requirement)
-        //{
-
-        //    if (requirement.timeout <= 0f)
-        //        requirement.timeout = float.MaxValue;
-
-        //    requirement.count = Math.Max(1, requirement.count);
-
-        //    return Listen(requirement.tag, Message_Analytics, requirement.timeout, requirement.count);
-        //}
-
+        
         /// <summary>
         /// Logs non-deferred text via either CyclopsCommon.Logger or UnityEngine.Debug.Log.
-        /// For lazy evaluation, please use: Log(Func<string> lazyPrinter)
+        /// For lazy evaluation, please use: Log(lazyPrinter)
         /// If CyclopsCommon.Logger is null, logging will default to UnityEngine.Debug.
         /// When using UnityEngine.Debug, logging is only available for debug builds.
         /// To log output without a Unity project reference, please ensure that CyclopsCommon.Logger refers to the desired logger.
@@ -176,14 +151,14 @@ namespace Smonch.CyclopsFramework
             if (Logger == null)
             {
 #if (UNITY_EDITOR || DEVELOPMENT_BUILD)
-                return Add(tag: Tag_Log, f: () => UnityEngine.Debug.Log(text));
+                return Add(tag: TagLog, f: () => UnityEngine.Debug.Log(text));
 #else
                 return Context;
 #endif
             }
             else
             {
-                return Add(tag: Tag_Log, f: () => Logger?.Invoke(text));
+                return Add(tag: TagLog, f: () => Logger?.Invoke(text));
             }
         }
 
@@ -193,21 +168,21 @@ namespace Smonch.CyclopsFramework
         /// When using UnityEngine.Debug, logging is only available for debug builds.
         /// To log output without a Unity project reference, please ensure that CyclopsCommon.Logger refers to the desired logger.
         /// </summary>
-        /// <param name="printer"></param>
+        /// <param name="lazyPrinter"></param>
         /// <returns></returns>
         public CyclopsRoutine Log(Func<string> lazyPrinter)
         {
             if (Logger == null)
             {
 #if (UNITY_EDITOR || DEVELOPMENT_BUILD)
-                return Add(tag: Tag_Log, f: () => UnityEngine.Debug.Log(lazyPrinter()));
+                return Add(tag: TagLog, f: () => UnityEngine.Debug.Log(lazyPrinter()));
 #else
                 return Context;
 #endif
             }
             else
             {
-                return Add(tag: Tag_Log, f: () => Logger?.Invoke(lazyPrinter()));
+                return Add(tag: TagLog, f: () => Logger?.Invoke(lazyPrinter()));
             }
         }
     }