+++ Runtime/Core/ICyclopsMessageInterceptor.cs
@@ -1,6 +1,6 @@
 // Cyclops Framework
 // 
-// Copyright 2010 - 2022 Mark Davis
+// Copyright 2010 - 2023 Mark Davis
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 namespace Smonch.CyclopsFramework
 {
-    interface ICyclopsMessageInterceptor
+    public interface ICyclopsMessageInterceptor
     {
         void InterceptMessage(CyclopsMessage msg);
     }