+++ Runtime/Game/CyclopsGame.cs
@@ -14,9 +14,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using System;
+using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Assertions;
 using UnityEngine.LowLevel;
+using Assert = UnityEngine.Assertions.Assert;
 
 #if UNITY_EDITOR
 using UnityEditor;
@@ -28,6 +30,8 @@ namespace Smonch.CyclopsFramework
     {
         private bool _isActive = false;
         private bool _isUsingAutomaticUpdateMode = false;
+        private PlayerLoopSystem _originalPlayerLoopSystem;
+        private Vector2Int _screenSize;
 
         public enum UpdateMode
         {
@@ -38,6 +42,13 @@ namespace Smonch.CyclopsFramework
         protected CyclopsStateMachine StateMachine { get; } = new();
         public bool IsQuitting { get; private set; }
 
+        public Action EarlyUpdateFinished { get; set; }
+        public Action FixedUpdateStarting { get; set; }
+        public Action FixedUpdateFinished { get; set; }
+        public Action LateUpdateStarting { get; set; }
+        public Action LateUpdateFinished { get; set; }
+        public Action ScreenSizeChanged { get; set; }
+        
         public CyclopsGame() { }
 
         public void Start(CyclopsBaseState initialState, UpdateMode updateMode)
@@ -52,23 +63,89 @@ namespace Smonch.CyclopsFramework
             if (updateMode == UpdateMode.Automatic)
                 _isUsingAutomaticUpdateMode = true;
 
+            _screenSize = new Vector2Int(Screen.width, Screen.height);
+
             StateMachine.PushState(initialState);
-
-            if (_isUsingAutomaticUpdateMode)
+            
+            _originalPlayerLoopSystem = PlayerLoop.GetCurrentPlayerLoop();
+            
+            PlayerLoopSystem cyclopsLoopSystem = new PlayerLoopSystem
             {
-                var currentPlayerLoop = PlayerLoop.GetCurrentPlayerLoop();
+                type = typeof(CyclopsGame),
+                updateDelegate = UpdateStateMachine
+            };
+            
+            PlayerLoopSystem earlyUpdateFinishedSystem = new PlayerLoopSystem
+            {
+                type = typeof(CyclopsGame),
+                updateDelegate = () => EarlyUpdateFinished?.Invoke()
+            };
+            
+            PlayerLoopSystem fixedUpdateStartingSystem = new PlayerLoopSystem
+            {
+                type = typeof(CyclopsGame),
+                updateDelegate = () => FixedUpdateStarting?.Invoke()
+            };
+            
+            PlayerLoopSystem fixedUpdateFinishedSystem = new PlayerLoopSystem
+            {
+                type = typeof(CyclopsGame),
+                updateDelegate = () => FixedUpdateFinished?.Invoke()
+            };
+            
+            PlayerLoopSystem lateUpdateStartingSystem = new PlayerLoopSystem
+            {
+                type = typeof(CyclopsGame),
+                updateDelegate = () => LateUpdateStarting?.Invoke()
+            };
+            
+            PlayerLoopSystem lateUpdateFinishedSystem = new PlayerLoopSystem
+            {
+                type = typeof(CyclopsGame),
+                updateDelegate = () => LateUpdateFinished?.Invoke()
+            };
+            
+            var currentPlayerLoop = PlayerLoop.GetCurrentPlayerLoop();
+            var existingSubsystems = currentPlayerLoop.subSystemList;
+            var modifiedSubsystems = new List<PlayerLoopSystem>(existingSubsystems.Length + 1);
 
-                for (int i = 0; i < currentPlayerLoop.subSystemList.Length; ++i)
+            foreach (var subsystem in existingSubsystems)
+            {
+                // Run after PlayerLoop.Update completes each frame.
+                if (_isUsingAutomaticUpdateMode && (subsystem.type == typeof(UnityEngine.PlayerLoop.Update)))
                 {
-                    if (currentPlayerLoop.subSystemList[i].type == typeof(UnityEngine.PlayerLoop.Update))
-                    {
-                        currentPlayerLoop.subSystemList[i].updateDelegate = UpdateStateMachine;
-                        break;
-                    }
+                    modifiedSubsystems.Add(subsystem);
+                    modifiedSubsystems.Add(cyclopsLoopSystem);
                 }
-                
-                PlayerLoop.SetPlayerLoop(currentPlayerLoop);
+                else if (subsystem.type == typeof(UnityEngine.PlayerLoop.EarlyUpdate))
+                {
+                    modifiedSubsystems.Add(subsystem);
+                    modifiedSubsystems.Add(earlyUpdateFinishedSystem);
+                }
+                else if (subsystem.type == typeof(UnityEngine.PlayerLoop.FixedUpdate))
+                {
+                    modifiedSubsystems.Add(fixedUpdateStartingSystem);
+                    modifiedSubsystems.Add(subsystem);
+                    modifiedSubsystems.Add(fixedUpdateFinishedSystem);
+                }
+                else if (subsystem.type == typeof(UnityEngine.PlayerLoop.PreLateUpdate))
+                {
+                    modifiedSubsystems.Add(subsystem);
+                    modifiedSubsystems.Add(lateUpdateStartingSystem);
+                }
+                else if (subsystem.type == typeof(UnityEngine.PlayerLoop.PostLateUpdate))
+                {
+                    modifiedSubsystems.Add(lateUpdateFinishedSystem);
+                    modifiedSubsystems.Add(subsystem);
+                }
+                else
+                {
+                    modifiedSubsystems.Add(subsystem);
+                }
             }
+            
+            currentPlayerLoop.subSystemList = modifiedSubsystems.ToArray();
+            PlayerLoop.SetPlayerLoop(currentPlayerLoop);
             
             Application.exitCancellationToken.Register(callback: Quit, useSynchronizationContext: true);
         }
@@ -95,25 +172,9 @@ namespace Smonch.CyclopsFramework
                 "Stop was already called and did not need to be called again.");
             
             StateMachine.ForceStop();
-            
             _isActive = false;
             
-            if (_isUsingAutomaticUpdateMode)
-            {
-                var currentPlayerLoop = PlayerLoop.GetCurrentPlayerLoop();
-
-                for (int i = 0; i < currentPlayerLoop.subSystemList.Length; ++i)
-                {
-                    if (currentPlayerLoop.subSystemList[i].type == typeof(UnityEngine.PlayerLoop.Update))
-                    {
-                        currentPlayerLoop.subSystemList[i].updateDelegate = null;
-                        break;
-                    }
-                }
-                
-                PlayerLoop.SetPlayerLoop(currentPlayerLoop);
-                _isUsingAutomaticUpdateMode = false;
-            }
+            PlayerLoop.SetPlayerLoop(_originalPlayerLoopSystem);
         }
 
         public void Update()
@@ -143,6 +204,15 @@ namespace Smonch.CyclopsFramework
             else
             {
                 StateMachine.Update();
+                
+                // Unity will not notify on screen size changes. This is useful enough to handle here.
+                var currentScreenSize = new Vector2Int(Screen.width, Screen.height);
+
+                if (currentScreenSize != _screenSize)
+                {
+                    _screenSize = currentScreenSize;
+                    ScreenSizeChanged?.Invoke();
+                }
             }
         }
     }